Fix bug causing certficates to be re-issued with misconfigured renewBefore/duration
