+++ pkg/controller/certificates/BUILD.bazel
@@ -54,8 +54,4 @@ go_test(
     name = "go_default_test",
     srcs = ["sync_test.go"],
     embed = [":go_default_library"],
-    deps = [
-        "//pkg/apis/certmanager/v1alpha1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
-    ],
 )
